requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
anyio==4.0.0
asttokens==2.4.1
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
executing==2.0.1
flaky==3.7.0
greenlet==3.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
icecream==2.1.3
idna==3.4
iniconfig==2.0.0
packaging==23.2
playwright==1.39.0
pluggy==1.3.0
pyee==11.0.1
Pygments==2.16.1
pytest==7.4.3
pytest-base-url==2.0.0
pytest-playwright==0.4.3
python-dotenv==1.0.0
python-slugify==8.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
termcolor==2.3.0
text-unidecode==1.3
typing_extensions==4.8.0
urllib3==2.1.0
